requirements.txt
langchain[all]
streamlit
// faiss-cpu
sentence_transformers
ctransformers
pandas
python-dotenv
tabulate
